aider: Update systray dependency to version 0.5.0.
# Aider chat conversation:

USER: i get the following error when running a cargo check `error: failed to select a version for the requirement `systray = "^0.5.1"``

ASSISTANT: The error message indicates that the version `0.5.1` of the `systray` crate is not available. This could be due to a typo in the version number or the version not being published on crates.io.

To resolve this issue, we can try to use the latest version of the `systray` crate. As of now, the latest version is `0.5.0`. Here is the proposed change:

```toml
Cargo.toml
<<<<<<< HEAD
[dependencies]
systray = "0.5.1"
=======
[dependencies]
systray = "0.5.0"
>>>>>>> updated
```

This change will update the `systray` dependency to version `0.5.0`.