aider: Added a new configuration in `.vscode/launch.json` to run the `main.py` file using Poetry in the virtual environment.
# Aider chat conversation:

USER: can this new task run it using poetry so the vertual environment is used?

ASSISTANT: Yes, to run the application using Poetry in the virtual environment, you need to modify the `.vscode/launch.json` file.

Here is the change:

```json
.vscode/launch.json
<<<<<<< HEAD
{
  "version": "0.2.0",
  "configurations": [
    {
      "name": "Python: Current File",
      "type": "python",
      "request": "launch",
      "program": "${workspaceFolder}/opener/main.py",
      "console": "integratedTerminal"
    }
  ]
}
=======
{
  "version": "0.2.0",
  "configurations": [
    {
      "name": "Python: Current File (Poetry)",
      "type": "python",
      "request": "launch",
      "program": "${workspaceFolder}/opener/main.py",
      "console": "integratedTerminal",
      "python": {
        "poetry": {
          "enabled": true
        }
      }
    }
  ]
}
>>>>>>> updated
```

This change adds a new configuration that uses Poetry to run the `main.py` file. The `"python.poetry.enabled": true` setting tells VSCode to use the Poetry virtual environment.

Please note that you need to have the Python extension for VSCode installed for this to work. If you don't have it installed, you can get it from the VSCode marketplace.