aider: Update Dependabot scanning to check for updates to Python/Poetry.
# Aider chat conversation:

USER: update the dependabot scanning to check for updates to python/poetry

ASSISTANT: To update the Dependabot scanning to check for updates to Python/Poetry, we need to modify the `.github/dependabot.yml` file. We will add a new update block for the Python package ecosystem and set the directory to the root of the project.

Here are the changes:

```yaml
.github/dependabot.yml
<<<<<<< HEAD
version: 2
updates:
  - package-ecosystem: "cargo" # See documentation for possible values
    directory: "/" # Location of package manifests
    schedule:
      interval: "weekly"
  - package-ecosystem: "github-actions" # See documentation for possible values
    directory: "/" # Location of package manifests
    schedule:
      interval: "weekly"
=======
version: 2
updates:
  - package-ecosystem: "cargo" # See documentation for possible values
    directory: "/" # Location of package manifests
    schedule:
      interval: "weekly"
  - package-ecosystem: "github-actions" # See documentation for possible values
    directory: "/" # Location of package manifests
    schedule:
      interval: "weekly"
  - package-ecosystem: "python" # See documentation for possible values
    directory: "/" # Location of package manifests
    schedule:
      interval: "weekly"
>>>>>>> updated
```

This will instruct Dependabot to check for updates to Python packages on a weekly basis.