aider: Added step to install semantic-release plugins before running the `cycjimmy/semantic-release-action` action.
# Aider chat conversation:

USER: I get the following error on the build for the Semantic Release step:Warning: The `set-output` command is deprecated and will be disabled soon. Please upgrade to using Environment Files. For more information see: https://github.blog/changelog/2022-10-11-github-actions-deprecating-save-state-and-set-output-commands/
Error: Error: Command failed: npm install @semantic-release/changelog @semantic-release/github --silent

ASSISTANT: The error message suggests that the `set-output` command is deprecated and will be disabled soon. This is a warning from GitHub Actions, and it's not the cause of the error. The actual error is the failure of the `npm install @semantic-release/changelog @semantic-release/github --silent` command.

This error is likely due to the `cycjimmy/semantic-release-action@v2` action not being able to install the necessary semantic-release plugins. This could be due to a number of reasons, such as network issues, npm registry issues, or issues with the action itself.

To fix this issue, you could try the following:

1. Use a specific version of the `cycjimmy/semantic-release-action` action instead of `v2`. This could potentially fix the issue if it's caused by a bug in the latest version of the action.

2. Install the semantic-release plugins manually before running the `cycjimmy/semantic-release-action` action. This could potentially fix the issue if it's caused by the action not being able to install the plugins.

Here is the edit block for the second solution:

```yaml
.github\workflows\build.yml
<<<<<<< HEAD
      - name: Semantic Release
        id: semantic
        uses: cycjimmy/semantic-release-action@v2
        with:
          semantic_version: 17
          branches: main
          extra_plugins: |
            @semantic-release/changelog
            @semantic-release/github
=======
      - name: Install semantic-release plugins
        run: |
          npm install @semantic-release/changelog @semantic-release/github
      - name: Semantic Release
        id: semantic
        uses: cycjimmy/semantic-release-action@v2
        with:
          semantic_version: 17
          branches: main
>>>>>>> updated
```

This edit block adds a new step to the workflow that installs the semantic-release plugins before running the `cycjimmy/semantic-release-action` action.