aider: Added a step to the build to cache the dependencies between runs.
# Aider chat conversation:

USER: add a step to the build to cache the dependancies between runs

ASSISTANT: To cache the dependencies between runs, we need to modify the `.github/workflows/build.yml` file. We will add a new step that uses the `actions/cache@v2` action to cache the Poetry dependencies. This step will be added after the "Set up Python" step and before the "Install dependencies" step.

Here is the edit block for the changes:

```yaml
.github/workflows/build.yml
<<<<<<< HEAD
    steps:
      - uses: actions/checkout@v2
      - name: Set up Python
        uses: actions/setup-python@v2
        with:
          python-version: "3.9"
      - name: Install dependencies
        run: |
          python -m pip install --upgrade pip
          pip install poetry
          poetry install
=======
    steps:
      - uses: actions/checkout@v2
      - name: Set up Python
        uses: actions/setup-python@v2
        with:
          python-version: "3.9"
      - name: Cache Poetry dependencies
        uses: actions/cache@v2
        with:
          path: ~/.cache/pypoetry
          key: ${{ runner.os }}-poetry-${{ hashFiles('**/poetry.lock') }}
          restore-keys: |
            ${{ runner.os }}-poetry-
      - name: Install dependencies
        run: |
          python -m pip install --upgrade pip
          pip install poetry
          poetry install
>>>>>>> updated
```

This new step will cache the Poetry dependencies in the `~/.cache/pypoetry` directory. The cache key is based on the runner's operating system and the hash of the `poetry.lock` file. If the `poetry.lock` file changes, a new cache will be created. If the `poetry.lock` file does not change, the action will restore the cache from a previous run.