aider: Updated the cache step in the build to also cache the pip packages.
# Aider chat conversation:

USER: update the cache step in the build to also cache the pip packages

ASSISTANT: To cache the pip packages, we need to modify the `.github/workflows/build.yml` file. We will add a new step using the `actions/cache@v2` action to cache the pip packages.

Here is the edit block for the changes:

```yaml
.github/workflows/build.yml
<<<<<<< HEAD
      - name: Cache Poetry dependencies
        uses: actions/cache@v2
        with:
          path: ~/.cache/pypoetry
          key: ${{ runner.os }}-poetry-${{ hashFiles('**/poetry.lock') }}
          restore-keys: |
            ${{ runner.os }}-poetry-
=======
      - name: Cache Poetry and pip dependencies
        uses: actions/cache@v2
        with:
          path: |
            ~/.cache/pypoetry
            ~\pip
          key: ${{ runner.os }}-poetry-pip-${{ hashFiles('**/poetry.lock') }}-${{ hashFiles('**/pip.lock') }}
          restore-keys: |
            ${{ runner.os }}-poetry-pip-
>>>>>>> updated
```

This change will cache both the Poetry and pip dependencies between runs, which should speed up the build process.